frontend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2 
supabase
dotenv